materials

+++ src/lighting/material.cpp
@@ -135,9 +135,9 @@ int main()
         cubeShader.setFloat("material.shininess",32.0f);
 
         glm::vec3 lightColor;
-        lightColor.x = glm::sin(glfwGetTime() * 2.0f);
-        lightColor.y = glm::sin(glfwGetTime() * 0.7f);
-        lightColor.z = glm::sin(glfwGetTime() * 1.3f);
+        lightColor.x = (float)glm::sin(glfwGetTime() * 2.0f);
+        lightColor.y = (float)glm::sin(glfwGetTime() * 0.7f);
+        lightColor.z = (float)glm::sin(glfwGetTime() * 1.3f);
         glm::vec3 ambientColor = lightColor * glm::vec3(0.2f);
         glm::vec3 diffuseColor = lightColor * glm::vec3(0.5f);
